+++ src/components/Button/Button.module.scss
@@ -1,4 +1,4 @@
-@use '../styles/color' as color;
+@use '../../assets/styles/color' as color;
 .button {
     /*box model*/
     height: 40px;
