+++ src/components/Main/TweetItemList/TweetItemList.module.scss
@@ -1,5 +1,5 @@
-@use '/src/components/styles/color';
-@use '/src/components/styles/mixin';
+@use '../../../assets/styles/color' as color;
+@use '../../../assets/styles/mixin';
 
 .itemContainer {
     display: flex;
@@ -99,7 +99,7 @@
             height: 16px;
             cursor: pointer;
             img {
-                content: url('../../../../assets/icons/tweet/tweet_reply.png');
+                content: url('../../../assets/icons/tweet/tweet_reply.png');
                 width: 100%;
                 display: block;
             }
@@ -120,7 +120,7 @@
         height: 16px;
         cursor: pointer;
         img {
-            content: url('../../../../assets/icons/tweet/tweet_like.png');
+            content: url('../../../assets/icons/tweet/tweet_like.png');
             width: 100%;
             display: block;
         }
