+++ src/components/SuggestUser/SuggestUserContainer/SuggestUserContainer.jsx
@@ -1,7 +1,7 @@
-import SuggestUserItem from 'components/UIComponents/SuggestUserItem/SuggestUserItem';
-import styles from './SuggestFollowContainer.module.scss';
+import SuggestUserItem from '../SuggestUserItem/SuggestUserItem.jsx';
+import styles from './SuggestUserContainer.module.scss';
 
-export default function SuggestFollowContainer() {
+export default function SuggestUserContainer() {
     return (
         <div className={styles.SuggestFollowContainer}>
             <h4 className={styles.suggestUserTitle}>推薦跟隨</h4>
