+++ src/components/SuggestUser/SuggestUserItem/SuggestUserItem.module.scss
@@ -1,5 +1,5 @@
-@use '../../styles/color' as color;
-@use '../../styles/mixin';
+@use '../../../assets/styles/color' as color;
+@use '../../../assets/styles/mixin';
 
 .userItem {
     width: 273px;
