+++ src/pages/ButtonShowCasePage/ButtonShowCasePage.jsx
@@ -1,10 +1,12 @@
 
 
 import Button from '../../components/Button/Button.jsx'
-import styles from './ButtonShowCasePage.module.scss'
+
 import TweetModal from '../../components/Modal/TweetModal/TweetModal.jsx'
 import ReplyModal from '../../components/Modal/ReplyModal/ReplyModal.jsx'
 // import UserEditModal from '../../components/Modal/UserEditModal/UserEditModal.jsx'
+
+import styles from './ButtonShowCasePage.module.scss'
 export default function ButtonShowCasePage () {
 
   return (
