+++ src/pages/CurrentUserTweetPage/CurrentUserTweetPage.module.scss
@@ -1,4 +1,4 @@
-@use '/src/components/styles/color' as color;
+@use '../../assets/styles/color' as color;
 
 .container {
     height: 100vh;
