+++ src/pages/MainPage/MainPage.jsx
@@ -1,11 +1,11 @@
 import styles from './MainPage.module.scss';
 
-import MainContainer from '../../components/MainContainer/MainContainer.jsx';
-import NavBarContainer from '../../components/NavBarContainer/NavBarContainer.jsx';
-import SuggestFollowContainer from '../../components/SuggestFollowContainer/SuggestFollowContainer.jsx';
+import MainContainer from '../../components/Main/MainContainer/MainContainer.jsx';
+import NavBarContainer from '../../components/Navbar/NavBarContainer/NavBarContainer.jsx';
+import SuggestUserContainer from '../../components/SuggestUser/SuggestUserContainer/SuggestUserContainer.jsx';
 import Header from '../../components/Header/Header.jsx';
-import TweetInput from '../../components/MainContainer/TweetInput/TweetInput.jsx';
-import TweetListItem from '../../components/UIComponents/NavItem/ListItem/TweetListItem.jsx';
+import TweetInput from '../../components/Main/TweetInput/TweetInput.jsx';
+import TweetItemList from '../../components/Main/TweetItemList/TweetItemList.jsx';
 
 export default function MainPage() {
     return (
@@ -17,11 +17,11 @@ export default function MainPage() {
             <MainContainer>
                 <Header title="首頁" />
                 <TweetInput />
-                <TweetListItem />
+                <TweetItemList />
             </MainContainer>
 
             {/* <div className={styles.suggestFollowContainer}> */}
-            <SuggestFollowContainer />
+            <SuggestUserContainer />
             {/* </div> */}
         </div>
     );
